Fix UnknownProperty being used for known properties
If the parser options IncludeUnknownDeclarations or AllowInvalidValues are set to true, StylesheetComposer will incorrectly use UnknownProperty for everything, even when the property name in question is a known one.

Fix by first trying to create the property with the specified name and only falling back to UnknownProperty if the created property is null and one of IncludeUnknownDeclarations  or AllowInvalidValues are set.