Fix busted tests due to a bug in PeriodicValueConverter
Some places was using just .Periodic() without any labels specified which triggered the debug assertion that the label count was 2 or 4. Because all the value converter stuff is static, the exception raised during the static constructor crashed the XUnit test runner process and masked the fact these tests were failing because they was never actually being run.

Fixed by treating the case of no labels the same as 4 for backwards compatibility.