Fix handling of unknown values
When the option to ignore unknown values was turned on, unknown values
were not being stored in the parsed properties as UnknownProperty was
only being used when an unknown property name was found, not an unknown
value for a known property.

To handle this, need to split the logic around unknown property into two
- firstly, if we seen an unknown property name then we instantiate an
UnknownProperty instance, but we if find a known property that we
can't set the value on (TrySetValue returns false), we need to convert
that property to an UnknownProperty instance and re-set the value on
it.