Fix @page rules with a selector not being added to the PageRules collection
When a @page rule had a selector (such as "@page :left", CreatePageSelector would consume the colon token for the pseudo-selector but NOT consume the pseudo-selector identifier token. This meant that when ParseComments was called after parsing the selector, the starting token it saw was an Ident token and so it skipped attempting to parse any comments or skip whitespace. The NextToken call after ParseComments would then potentially leave the current token as a whitespace token, rather than the opening curly-brace to start the declarations and so no declarations would be parsed and no page rule added to the object model.

Fixed by consuming the pseudo-selector identifier token in CreatePageSelector so that the call to ParseComments will skip over any comments and whitespace. As a result, ParseComments will now exit when the current token is the opening curly brace so the NextToken call after ParseComments is not needed.